backend/test_models.py
import os
import google.generativeai as genai
from dotenv import load_dotenv

load_dotenv()
api_key = os.getenv("GEMINI_API_KEY")

with open('models.txt', 'w') as f:
    if not api_key or api_key == "your_gemini_api_key_here":
        f.write("Please set your GEMINI_API_KEY in .env!\n")
    else:
        genai.configure(api_key=api_key)
        f.write("Available Text Models:\n")
        try:
            for m in genai.list_models():
                if 'generateContent' in m.supported_generation_methods:
                    f.write(f"{m.name}\n")
        except Exception as e:
            f.write(f"Error checking models: {e}\n")
