backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
scikit-learn
numpy
pandas
tensorflow
Pillow
httpx
pydantic
joblib
google-generativeai
python-dotenv
